Add support for sending Bter UI frames at lapdm.c
Bter frames may be used on downlink of a main DCCH or SACCH channel.
It applies to SAPI 0 and UI frames only. It includes a short layer 2
header with 2 bits only. All other bits are used for the message. The
message size is 23 bytes on main DCCH and 21 bytes on SACCH.
The length of the L3 messsage is used to distinguish between Bter frames
and other (UI) frames.

Note that the L3 message header is different, so that the length of the
UI frame will determine which header is used.

Change-Id: Ia3a25c009d1ff09f83258bdb226a85b81466d7a1