struct osmo_sub_auth_data: remove OSMO_DEPRECATED_OUTSIDE
I was the one who suggested adding this attribute during code review,
and now, having realized it was a bad idea, I am removing it.  The
problem is that this attribute spams compilation logs of projects
including file <osmocom/crypt/auth.h>, even if the deprecated struct
is not used directly at all.

Change-Id: Ia5d365206207872d5d3fdd4ae40273eab909fb33
Fixes: 08450c9e "libosmogsm: Support authentication with 256-bit K and/or OP/OPc"