lapdm: Do not return an error when enqueuing a frame
If tx_ph_data_enqueue() is called, frames will be written into a queue,
if there is a pending frame or if polling of TX frames is used. In
this case the return value must be 0.

Sending a RSL_MT_UNIT_DATA_REQ from upper layer causes a call to
tx_ph_data_enqueue(). The return code is returned to the sender. The
sender must not get an error returned, if the message is enqueued.

Change-Id: Iaeaf7c66cb3cf5cc81bc8e15d468e8e7704c1407