sockaddr_str.h Fix OSMO_SOCKADDR_STR_FMT_ARGS_NOT_NULL syntax error
It's fine changing the define because due to the previous error it means
basically that it was never used so far, because using it triggers a
compilation error.

The error was introduced because I ended up not using this in the past
after submitting this code, and now that I want to use it it went
noticed that it fails.

Change-Id: Iee361d740845257fa62c9093e30e8079fa933827