soft_uart: check Rx/Tx state once in osmo_soft_uart_{rx,tx}_ubits()
Check it once rather than doing this in a loop.  Return -EAGAIN if
Rx or Tx is not enabled when calling osmo_soft_uart_{rx,tx}_ubits().

This [theoretically] improves performance by reducing the number of
conditional statements in loops.  In the Tx path, this also prevents
calling the .tx_cb() when the transmitter is disabled, so that we
don't loose the application data.

Change-Id: I70f93b3655eb21c2323e451052c40cd305c016c8
Related: OS#4396