WIP: migrate gsmtap logging over to osmo_io
This is working in principle;  The problem is just that osmo_io itself
contains LOGP calls from within osmo_iofd_write_msgb() and downstream
code, so we might easily get up in an infinite recursion bug.

This needs further study; possibly we need a "safe" logging function
which will never use osmo_io.  Or something like a per-thread global
boolean variable which the code can check to know if it's re-entering?

Change-Id: I69d7eca7d50666b4fcfe73a1dc03f8e8491b32fc