osmo_io: Increase default queue max_length from 32 to 1024
The default value is way too low, and can be hit easily by apps
submitting several messages in one main loop run.
This is the case for instance of the Abis OML protocol, where a cascade
of "State Changed Event Reports" are sent for the whole object tree.

Set the default to a saner bigger value. In the end, increasing the
value doesn't hurt normal use of apps, and 1024 is still a fairly small
number to avoid the app eating the whole system memory if something
unexpected happens.

Change-Id: I92d5369af523f6297d45d4b534bb0f349ae74485