gprs_ns2_sns: improve search for same NS-VC
Adding a second NS-VC with the same IP+UDP isn't possible,
even when the weights are different, because the NS-ALIVE wouldn't work.
Also a NS-VC is identified by IP+UDP and not by the weights.

IPv6 did this already. Use the same logic for IPv4.
Further relax the search of NS-VC when removing NS-VC. Only take care
of IP+UDP and ignore signalling weight and data weights.

Related: OS#6611
Change-Id: Ibbc1a794a9a9f0ce6e583fd9743e7f02d52a99cf