tests/osmo_io: Rework test_file()
The amount of read and write on each callback is actually not
really managed by us but by the kernel. Instead of validating the amount
of read/write callbacks, monitor the amount of read/write bytes (and
also make sure the received a final read with len=0 to signal EOF).

Add fflush(stdout) before calling ASSERT to make sure stdout is flushed
before terminating the program.

Related: OS#6708
Change-Id: Ie1e686b407532eee76a8692dfc632af73183e614