iuup: Explicitly discard Iu-UP-DATA.req in Initialization state
Since Initialization procedure is asynchronous (need to wait for IuUP
ACK from peer), it may be that the user initializes the IuUP layer and
starts sending/forarding IuUP data to it before the initialization
finishes.
Before this patch, an error would be logged for each such Data request.
Let's instead explicitly expect this event and hence avoid logging a
generic FSM error log "{Initialisation}: Event IuUP-DATA-req not
permitted" and log in INFO level and discard the message.

Change-Id: I424be5718cdd12cfd0ce80108a002ce68d3e17aa