configure: set min python version check to 3.4
While the osmo-gsm-tester builder jobs that use the poky SDK with python
version 3.5 are passing with the last patch, the jobs that build our
images and packages with the poky SDK (201705 jobs) are still failing.
This is because they run on a machine with Python 3.4 installed, and
that version gets used in the build process. Lower the version to 3.4 to
not have a regression there.

Fixes: 489a6f ("configure.ac: check for Python interpreter unconditionally")
Fixes: 40e8f5 ("configure: set min python version check to 3.5")
Change-Id: Ia83f360274061463232c1fcd7f6dc6e70f774738