Makefile: fix doc/vty race condition
It is possible that doxygen is unable to create the "doc/vty" directory,
as we have seen in jenkins. Create it before running doxygen as it is
done for other doc dirs above, e.g.:

  rm -rf $(top_builddir)/doc/isdn; mkdir -p $(top_builddir)/doc/isdn

Note that the doc/vty dir is expected to exist already inside the source
tree, as it contains versioned failes. As I understand it, the error
happens when building out-of-tree.

Fix for:
  error: tag OUTPUT_DIRECTORY: Output directory 'doc/vty' does not exist and cannot be created
  Doxygen version used: 1.9.4
  Exiting...
  touch "doc/libosmovty.tag.prep"
  make[2]: [Makefile:1357: doc/libosmovty.tag.prep] Error 1 (ignored)
  touch: cannot touch 'doc/libosmovty.tag.prep': No such file or directory

Related: https://gerrit.osmocom.org/c/libosmocore/+/39827/1#message-02393b0dc4323a19eb2fd2a19422f50439b81953
Change-Id: If9029059ac60060bdb14f5629f50836166e3afd7