osmo-config-merge: Fix comments with different indentation creating merge problems
VTY comments are not part of the tree structure, since they can be
placed anywhere and are just ignored.
Hence, ignore them when generating the tree structure during file read.

This has the backside that of course when merging files, comments on the
0-level show up in the end in testB.ok since the content below it in
testB.confmerge is now merged (properly) into previous nodes.
This can be slightly improved by placing the comments in the proper
indentation level, since if a new node is appended from the .confmerge,
they will be added in order and hence will show up properly.

Change-Id: I7de78d3ed3ca27e3470ea66a78ff9c5acfe65741