osmo_io: Keep msgb ownership internal during segmentation until read_cb
Until dispatched to the user over read_cb(), the msgb is considered to
be not-ready-for-user and hence it's kept owned by the iofd.
This resembles more the actual expected parentship tree, since
iofd->pending is freed by osmo_io internally in osmo_iofd_close().

More importat, which should cause less trouble if for instance user of
osmo_iofd wishes to reparent the osmo_io and then free the previous ctx
it passed during osmo_iofd_setup().

Change-Id: I4f41153948b7f1568c7499db53b9620bf2dd9ac4