Add multiple messages buffers to struct iofd_msghdr
This is a prerequisite for transfering multiple buffers with a single
SQE or CQE when using osmo_io with io_uring.

The message buffers at struct iofd_msghr, as well as the io vectors are
arrays. A maximum number is stored, as well as the current number.

As this patch does not implement multiple buffers yet, only index 0 of a
message buffer is used. (msg[0])

Having more than one buffer is optional and the number can be controlled
via environment variable.

Related: OS#6705
Change-Id: I4fb1067de4615cc22cc6caf99b481491e7f2ef92