AMR codec: add conversion between s-bits and codec parameters
At the lowest level of frame information content, before introducing
higher layers of RTP or TRAU frame or other encoding, an AMR codec
frame (speech or SID) can be represented in 3 fundamental ways:

* 35 to 244 s-bits as defined theoretically in TS 26.090 & 26.092
  and used in practice in 2G TRAU frames;

* 39 to 244 d-bits as defined theoretically in TS 26.101 and used in
  practice in 3G IuUP and in RTP;

* 5 to 57 codec parameters, each fitting into uint16_t, as implemented
  in 3GPP reference C code (TS 26.073) and assumed by the same spec
  in its listing of decoder homing frame bit patterns.

libosmocodec previously provided conversion functions between s-bits
and d-bits, as well as a table that gives the number of d-bits per
mode - but no conversion to or from arrays of codec parameters.
Add the latter functionality.

Change-Id: I28eae702204d7f7ff164e17c5e26152a94a12cba