gsm/protocol/gsm_48_103.h: add definition for TW-TS-007
TW-TS-007 is a Themyscira/Osmocom extension to AoIP interface
that provides a much more compact payload format for CSD, matching
the size of compressed speech payloads.  Clause 9 in this spec
fixes the payload type on AoIP interface for this pseudo-codec
as 127, distinguishing it from CLEARMODE and from all speech
codecs.

Osmocom applications that may need payload type numbers for the
standard set of codecs defined in TS 48.103 may also need the
payload type for TW-TS-007 if they implement this extension,
as this payload format is intended to "stand on equal footing"
with standard AoIP payload types.  Therefore, add it to the same
header file.

Change-Id: I74c9e76ebcce32ad20501134ec5089794acb4f62