gsm/rsl: add definition for TFO transparent container IE
3GPP Rel5 version of TS 48.058 RSL spec introduced a new IE named
"TFO transparent container".  By this point in GSM/3GPP evolutionary
history, Abis was already treated as a de facto proprietary interface,
and the spec is silent regarding the exact bit content of this new IE.
(See vague description in section 9.3.59.)  However, this IE was added
at the same time as Generic Config Frames for TFO (TS 28.062 Annex H),
hence one can reasonably infer that this IE is intended to transport
the content of these GCFs between the BTS and the BSC, with the BTS
relaying, mostly transparently, between the BSC and the TRAU channel.

Add RSL_IE_TFO_XPAR_CONT definition, to make it possible to add this
IE to CHAN ACTIV and MODE MODIFY commands (and thus enable TFO-AMR
on those E1 BTS models that implement Rel5 version thereof) and to
parse BTS response and notification messages that may also contain
this IE.

Change-Id: Ic6839827fb2bf8e401476af1d7907b577e5f094d